Refactor benchmarks.
