+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import xml.etree.ElementTree as ET
 from enum import Enum
 
 import pytest
@@ -15,6 +14,7 @@ def pytest_make_parametrize_id(config, val, argname):
 
 
 def pytest_addoption(parser):
+    parser.addoption("--backend", action="store", default=gs.cpu, help="Default simulation backend.")
     parser.addoption("--vis", action="store_true", default=False, help="Enable interactive viewer.")
 
 
@@ -26,14 +26,23 @@ def show_viewer(pytestconfig):
 @pytest.fixture
 def backend(request):
     if hasattr(request, "param"):
-        return getattr(gs.constants.backend, "cpu")
-    return gs.cpu
+        backend = request.param
+        if isinstance(backend, str):
+            return getattr(gs.constants.backend, backend)
+        return backend
+    return pytestconfig.getoption("--backend")
 
 
 @pytest.fixture(scope="function", autouse=True)
 def initialize_genesis(request, backend):
     logging_level = request.config.getoption("--log-cli-level")
-    gs.init(backend=backend, precision="64", logging_level=logging_level, seed=0, debug=True)
+    if backend == gs.cpu:
+        precision = "64"
+        debug = True
+    else:
+        precision = "32"
+        debug = False
+    gs.init(backend=backend, precision=precision, debug=debug, seed=0, logging_level=logging_level)
     yield
     gs.destroy()
 
@@ -99,27 +108,3 @@ def gs_sim(xml_path, gs_solver, gs_integrator, show_viewer, mj_sim):
     gs_sim.scene.build()
 
     return gs_sim.scene.sim
-
-
-@pytest.fixture
-def xml_path(request, tmp_path, model_name):
-    mjcf = request.getfixturevalue(model_name)
-    xml_tree = ET.ElementTree(mjcf)
-    file_path = tmp_path / f"{model_name}.xml"
-    xml_tree.write(file_path, encoding="utf-8", xml_declaration=True)
-    return str(file_path)
-
-
-@pytest.fixture(scope="session")
-def box_plan():
-    """Generate an XML model for a box on a plane."""
-    mjcf = ET.Element("mujoco", model="one_box")
-    ET.SubElement(mjcf, "option", timestep="0.01")
-    default = ET.SubElement(mjcf, "default")
-    ET.SubElement(default, "geom", contype="1", conaffinity="1", condim="3", friction="1. 0.5 0.5")
-    worldbody = ET.SubElement(mjcf, "worldbody")
-    ET.SubElement(worldbody, "geom", type="plane", name="floor", pos="0. 0. 0.", size="40. 40. 40.")
-    box_body = ET.SubElement(worldbody, "body", pos="0. 0. 0.3")
-    ET.SubElement(box_body, "geom", type="box", size="0.2 0.2 0.2", pos="0. 0. 0.")
-    ET.SubElement(box_body, "joint", type="free")
-    return mjcf
