+++ tests/test_rigid_physics.py
@@ -1,0 +1,45 @@
+import pytest
+import xml.etree.ElementTree as ET
+
+import numpy as np
+import genesis as gs
+
+from .utils import simulate_and_check_mujoco_consistency
+
+
+@pytest.fixture
+def xml_path(request, tmp_path, model_name):
+    mjcf = request.getfixturevalue(model_name)
+    xml_tree = ET.ElementTree(mjcf)
+    file_path = tmp_path / f"{model_name}.xml"
+    xml_tree.write(file_path, encoding="utf-8", xml_declaration=True)
+    return str(file_path)
+
+
+@pytest.fixture(scope="session")
+def box_plan():
+    """Generate an XML model for a box on a plane."""
+    mjcf = ET.Element("mujoco", model="one_box")
+    ET.SubElement(mjcf, "option", timestep="0.01")
+    default = ET.SubElement(mjcf, "default")
+    ET.SubElement(default, "geom", contype="1", conaffinity="1", condim="3", friction="1. 0.5 0.5")
+    worldbody = ET.SubElement(mjcf, "worldbody")
+    ET.SubElement(worldbody, "geom", type="plane", name="floor", pos="0. 0. 0.", size="40. 40. 40.")
+    box_body = ET.SubElement(worldbody, "body", pos="0. 0. 0.3")
+    ET.SubElement(box_body, "geom", type="box", size="0.2 0.2 0.2", pos="0. 0. 0.")
+    ET.SubElement(box_body, "joint", type="free")
+    return mjcf
+
+
+@pytest.mark.parametrize("model_name", ["box_plan"])
+@pytest.mark.parametrize("gs_solver", [gs.constraint_solver.CG])
+@pytest.mark.parametrize("gs_integrator", [gs.integrator.implicitfast])
+@pytest.mark.parametrize("backend", ["cpu"], indirect=True)
+def test_box_plan_dynamics(gs_sim, mj_sim):
+    (gs_robot,) = gs_sim.entities
+    cube_pos = np.array([0.0, 0.0, 0.6])
+    cube_quat = np.random.rand(4)
+    cube_quat /= np.linalg.norm(cube_quat)
+    qpos = np.concatenate((cube_pos, cube_quat))
+    qvel = np.zeros((gs_robot.n_dofs,))
+    simulate_and_check_mujoco_consistency(gs_sim, mj_sim, qpos, qvel, num_steps=150)
